mary: summarizes the given text
Mary is a Perl script that reads a given text,
find the most frequent word and summarizes using
it in combination with conjunctions, prepositions,
articles and verbs.

It was programmed to use Portuguese as the
working language, but can easily be changed
to work with English, for example.

Usage:

perl mary.pl <text_to_summarize>

The output will be put in a file called 'resumo'.